+++ MD5.h
@@ -27,7 +27,7 @@
 
 #include <string.h>
 
-typedef unsigned long MD5_u32plus;
+typedef uint32_t MD5_u32plus;
 
 typedef struct {
 	MD5_u32plus lo, hi;
@@ -42,7 +42,7 @@ public:
 	MD5();
 	static unsigned char* make_hash(char *arg);
 	static unsigned char* make_hash(char *arg,size_t size);
-	static char* make_digest(const unsigned char *digest, int len);
+	static char* make_digest(const unsigned char *digest, int16_t len);
  	static const void *body(void *ctxBuf, const void *data, size_t size);
 	static void MD5Init(void *ctxBuf);
 	static void MD5Final(unsigned char *result, void *ctxBuf);
